meme/requirements.txt
Flask
Pillow
mediapipe
numpy<2.0
